Implement -addChild: and -removeChild: as conveniences on HTMLNode.
Turns out adding an object already in the ordered set calls the KVC method -insertObject:inChildrenAtIndex: with an index one beyond the last index. An NSMutableOrderedSet is documented to do nothing when -addObject: is called with an already-present object. However, by blindly passing along -insertObject:atIndex: to the underlying NSMutableOrderedSet, the object was silently *removed* from the set of children. So now we check for the object's presence and correctly do nothing when it's already present.

Closes #57.