Update encoding list from the WHATWG and fix name lookups.
The changes are the addition of some new values and some capitalization of canonical names.

The capitalization changes pointed out that we have no reason to do a case-insensitive compare of the canonical names. In fact, since we generate the sorted list by doing a case-sensitive compare, this was incorrect code. However, previously all the names were lowercase, so I don't believe the incorrect code returned incorrect results.

The canonical names are internal to HTMLReader, they are not returned by any public API.